pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "veridex"
version = "0.2.0"
description = "A modular, probabilistic, and research-grounded AI content detection library."
readme = "README.md"
authors = [
  { name = "Veridex Contributors" },
]
license = { file = "LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
text = [
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "nltk>=3.8"
]
image = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "diffusers>=0.20.0",
    "Pillow>=9.0.0",
    "opencv-python-headless>=4.0.0"
]
audio = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0"
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0"
]
video = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python-headless>=4.0.0",
    "scipy>=1.7.0",
    "librosa>=0.10.0",
    "mediapipe>=0.10.0" 
]

[project.urls]
"Homepage" = "https://github.com/ADITYAMAHAKALI/veridex"
"Bug Tracker" = "https://github.com/ADITYAMAHAKALI/veridex/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["veridex*"]
